--- a/hw10_program_optimization/new.txt
+++ b/hw10_program_optimization/new.txt
@@ -2,6 +2,6 @@ goos: linux
 goarch: amd64
 pkg: github.com/fixme_my_friend/hw10_program_optimization
 cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
-Benchmark_GetDomainStat-8   	1000000000	         0.0000244 ns/op	       0 B/op	       0 allocs/op
+Benchmark_GetDomainStat-8   	       8	 140189349 ns/op	 3365291 B/op	  123797 allocs/op
 PASS
-ok  	github.com/fixme_my_friend/hw10_program_optimization	0.004s
+ok  	github.com/fixme_my_friend/hw10_program_optimization	2.242s