--- a/hw10_program_optimization/old.txt
+++ b/hw10_program_optimization/old.txt
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/fixme_my_friend/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
+Benchmark_GetDomainStat-8   	1000000000	         0.2543 ns/op	       0 B/op	       0 allocs/op
+PASS
+ok  	github.com/fixme_my_friend/hw10_program_optimization	3.704s