--- a/hw10_program_optimization/old.txt
+++ b/hw10_program_optimization/old.txt
@@ -2,6 +2,6 @@ goos: linux
 goarch: amd64
 pkg: github.com/fixme_my_friend/hw10_program_optimization
 cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
-Benchmark_GetDomainStat-8   	1000000000	         0.2543 ns/op	       0 B/op	       0 allocs/op
+Benchmark_GetDomainStat-8   	       2	 663750516 ns/op	292574508 B/op	 2345821 allocs/op
 PASS
-ok  	github.com/fixme_my_friend/hw10_program_optimization	3.704s
+ok  	github.com/fixme_my_friend/hw10_program_optimization	2.009s