--- a/hw10_program_optimization/stats_bench_test.go
+++ b/hw10_program_optimization/stats_bench_test.go
@@ -4,18 +4,28 @@
 package hw10programoptimization
 
 import (
-	"bytes"
+	"archive/zip"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
-func Benchmark_GetDomainStat(b *testing.B) {
-	data := []byte(`{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
-{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
-{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
-{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
-{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`)
+// go test -bench=. -benchmem  > name.txt -tags=bench -run=BenchmarkGetDomainStat
+// benchcmp old.txt new.txt
 
-	if _, err := GetDomainStat(bytes.NewReader(data), "com"); err != nil {
-		b.Fatal(err)
+func Benchmark_GetDomainStat(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		r, err := zip.OpenReader("testdata/users.dat.zip")
+		require.NoError(b, err)
+		defer r.Close()
+
+		require.Equal(b, 1, len(r.File))
+
+		data, err := r.File[0].Open()
+		require.NoError(b, err)
+
+		if _, err := GetDomainStat(data, "com"); err != nil {
+			b.Fatal(err)
+		}
 	}
 }